refactor: move inference helpers into engine
